==> 00-generate_daily_weather.ipynb <==
import pandas as pd

df1 = pd.read_csv('data/weather/weather_merignac_daily_raw_1.csv', sep=';', decimal=',')
print('Proportion of missing lines:')
display(df1.isnull().sum()/len(df1))

<table class="liste" width="75%" cellspacing="2" cellpadding="2" border="1"><tbody>
<tr><th>Mnémonique</th><th>Libellé</th><th>Unité</th><th>Pas de temps</th></tr>
<tr><td>RR</td><td>HAUTEUR DE PRECIPITATIONS QUOTIDIENNE</td><td>MILLIMETRES&nbsp;ET&nbsp;1/10</td><td>quotidien</td></tr>
<tr><td>TN</td><td>TEMPERATURE MINIMALE SOUS ABRI QUOTIDIENNE</td><td>DEG&nbsp;C&nbsp;ET&nbsp;1/10</td><td>quotidien</td></tr>
<tr><td>TX</td><td>TEMPERATURE MAXIMALE SOUS ABRI QUOTIDIENNE</td><td>DEG&nbsp;C&nbsp;ET&nbsp;1/10</td><td>quotidien</td></tr>
<tr><td>TM</td><td>TEMPERATURE MOYENNE SOUS ABRI QUOTIDIENNE</td><td>DEG&nbsp;C&nbsp;ET&nbsp;1/10</td><td>quotidien</td></tr>
<tr><td>PMERM</td><td>PRESSION MER MOYENNE QUOTIDIENNE</td><td>HPA&nbsp;ET&nbsp;1/10</td><td>quotidien</td></tr>
<tr><td>FF2M</td><td>MOYENNE DES VITESSES DU VENT A 2 METRES QUOTIDIENNE</td><td>M/S&nbsp;ET&nbsp;1/10</td><td>quotidien</td></tr>
<tr><td>INST</td><td>DUREE D'INSOLATION QUOTIDIENNE</td><td>MINUTES</td><td>quotidien</td></tr>
<tr><td>GLOT</td><td>RAYONNEMENT GLOBAL QUOTIDIEN</td><td>JOULES/CM2</td><td>quotidien</td></tr>
<tr><td>GRELE</td><td>OCCURRENCE DE GRELE QUOTIDIENNE</td><td>BOOLEEN</td><td>quotidien</td></tr>
<tr><td>ETPMON</td><td>EVAPO—TRANSPIRATION MONTEITH QUOTIDIENNE</td><td>MILLIMETRES&nbsp;ET&nbsp;1/10</td><td>quotidien</td></tr>
</tbody></table>

df1[df1['GRELE'].isnull()].tail(20)

df2 = pd.read_csv('data/weather/weather_merignac_daily_raw_2.csv', sep=';', decimal=',')
print('Proportion of missing lines:')
display(df2.isnull().sum()/len(df2))

<table class="liste" width="75%" cellspacing="2" cellpadding="2" border="1"><tbody>
<tr><th>Mnémonique</th><th>Libellé</th><th>Unité</th><th>Pas de temps</th></tr>
<tr><td>FFM</td><td>MOYENNE DES VITESSES DU VENT A 10M QUOTIDIENNE</td><td>M/S&nbsp;ET&nbsp;1/10</td><td>quotidien</td></tr>
<tr><td>DIRT</td><td>RAYONNEMENT&nbsp;DIRECT&nbsp;QUOTIDIEN</td><td>JOULES/CM2</td><td>quotidien</td></tr>
</tbody></table>

df = df1.merge(df2, on = ['POSTE', 'DATE']).drop(['POSTE','FF2M', 'GLOT', 'DIRT'], axis=1)

df['Date'] = pd.to_datetime(df['DATE'].astype(str))
df = df.set_index('Date').drop('DATE', axis=1).rename(
    columns={
        'RR':'P',
        'TN':'Tn',
        'TX':'Tx',
        'TM':'Tm',
        'PMERM':'Pr',
        'INST':'Sun',
        'GRELE':'Hail',
        'ETPMON':'ETPmon',
        'FFM':'Wind'
    })

# Output to csv

df.to_csv('data/weather_merignac_daily.csv')

# SAFRAN Model

  <table class="liste" width="75%" cellspacing="2" cellpadding="2" border="1"><tbody>
<tr><th>Mnémonique</th><th>Libellé</th><th>Unité</th><th>Pas de temps</th></tr>
<tr><td>PRENEI_Q</td><td>Précipitations solides, cumul quotidien, SAFRAN</td><td>mm et dixièmes</td><td>quotidien</td></tr>
<tr><td>PRELIQ_Q</td><td>Précipitations liquides, cumul quotidien, SAFRAN</td><td>mm et dixièmes</td><td>quotidien</td></tr>
<tr><td>T_Q</td><td>Température, moyenne quotidienne, SAFRAN</td><td>°C et dixièmes</td><td>quotidien</td></tr>
<tr><td>FF_Q</td><td>Vent, moyenne quotidienne, SAFRAN</td><td>m/s et dixièmes</td><td>quotidien</td></tr>
<tr><td>SSI_Q</td><td>Rayonnement visible, cumul quotidien, SAFRAN</td><td>J/cm²</td><td>quotidien</td></tr>
<tr><td>EVAP_Q</td><td>Evaporation totale réelle, cumul quotidien, SURFEX</td><td>mm et dixièmes</td><td>quotidien</td></tr>
<tr><td>ETP_Q</td><td>Evapotranspiration potentielle (formule de Penman—Monteith)</td><td>mm et dixièmes</td><td>quotidien</td></tr>
<tr><td>PE_Q</td><td>Pluies efficaces, cumul quotidien</td><td>mm et dixièmes</td><td>quotidien</td></tr>
<tr><td>SWI_Q</td><td>Indice d humidité des sols, moyenne quotidienne, SURFEX</td><td> </td><td>quotidien</td></tr>
<tr><td>TINF_H_Q</td><td>Température minimale des 24 températures horaires, SAFRAN</td><td>°C et dixièmes</td><td>quotidien</td></tr>
<tr><td>TSUP_H_Q</td><td>Température maximale des 24 températures horaires, SAFRAN</td><td>°C et dixièmes</td><td>quotidien</td></tr>
</tbody></table>

#### St Emilion

Références géographiques

Point N° 7256 (Lon : —0.145  / Lat : 44.91 )

df_st_em = pd.read_excel("data/weather/weather_st_emilion_daily_raw.xls").drop(
    ["NUMERO", "LONGITUDE", "LATITUDE"], axis=1
)
df_st_em_additional = pd.read_excel("data/weather/weather_st_emilion_additional_daily_raw.xls").drop(
    ["NUMERO", "LAMBERTX", "LAMBERTY"], axis=1
)
df_st_em = pd.concat([df_st_em, df_st_em_additional])
df_st_em["DATE"] = df_st_em["DATE"].dt.normalize()
df_st_em = df_st_em[df_st_em["DATE"].dt.year >= 1960]
df_st_em = df_st_em.rename(
    columns={
        "PRENEI_Q": "Ps",
        "PRELIQ_Q": "Pl",
        "T_Q": "Tm",
        "FF_Q": "Wind",
        "SSI_Q": "Sun",
        "EVAP_Q": "Evap",
        "ETP_Q": "ETP",
        "PE_Q": "Peff",
        "SWI_Q": "SoilHum",
        "TINF_H_Q": "Tn",
        "TSUP_H_Q": "Tx",
        "DATE": "Date",
    }
).set_index("Date")
df_st_em["P"] = df_st_em["Ps"] + df_st_em["Pl"]
display(df_st_em.head())
df_st_em.info()

df_st_em.to_csv("data/weather/generated_weather_st_em.csv")

#### Pauillac

Point N° 6931 (X : 3560 hm / Y : 20250 hm)

LON = -0.767, LAT = 45.182

df_pauillac = pd.read_excel("data/weather/weather_pauillac_daily_raw.xls").drop(
    ["NUMERO", "LAMBERTX", "LAMBERTY"], axis=1
)
df_pauillac_additional = pd.read_excel("data/weather/weather_pauillac_additional_daily_raw.xls").drop(
    ["NUMERO", "LAMBERTX", "LAMBERTY"], axis=1
)
df_pauillac = pd.concat([df_pauillac, df_pauillac_additional])
df_pauillac["DATE"] = df_pauillac["DATE"].dt.normalize()
df_pauillac = df_pauillac[df_pauillac["DATE"].dt.year >= 1960]
df_pauillac = df_pauillac.rename(
    columns={
        "PRENEI_Q": "Ps",
        "PRELIQ_Q": "Pl",
        "T_Q": "Tm",
        "FF_Q": "Wind",
        "SSI_Q": "Sun",
        "EVAP_Q": "Evap",
        "ETP_Q": "ETP",
        "PE_Q": "Peff",
        "SWI_Q": "SoilHum",
        "TINF_H_Q": "Tn",
        "TSUP_H_Q": "Tx",
        "DATE": "Date",
    }
).set_index("Date")
df_pauillac["P"] = df_pauillac["Ps"] + df_pauillac["Pl"]
display(df_pauillac.head())
df_pauillac.info()

df_pauillac.to_csv("data/weather/generated_weather_pauillac.csv")

#### Leognan
Point N 7418 (Lon : -0.542 / Lat: 44.756)

df_leognan = pd.read_excel("data/weather/weather_leognan_daily_raw.xls").drop(
    ["NUMERO", "LONGITUDE", "LATITUDE"], axis=1
)
df_leognan_additional = pd.read_excel("data/weather/weather_leognan_additional_daily_raw.xls").drop(
    ["NUMERO", "LAMBERTX", "LAMBERTY"], axis=1
)
df_leognan = pd.concat([df_leognan, df_leognan_additional])
df_leognan["DATE"] = df_leognan["DATE"].dt.normalize()
df_leognan = df_leognan[df_leognan["DATE"].dt.year >= 1960]
df_leognan = df_leognan.rename(
    columns={
        "PRENEI_Q": "Ps",
        "PRELIQ_Q": "Pl",
        "T_Q": "Tm",
        "FF_Q": "Wind",
        "SSI_Q": "Sun",
        "EVAP_Q": "Evap",
        "ETP_Q": "ETP",
        "PE_Q": "Peff",
        "SWI_Q": "SoilHum",
        "TINF_H_Q": "Tn",
        "TSUP_H_Q": "Tx",
        "DATE": "Date",
    }
).set_index("Date")
df_leognan["P"] = df_leognan["Ps"] + df_leognan["Pl"]
display(df_leognan.head())
df_leognan.info()

df_leognan.to_csv("data/weather/generated_weather_leognan.csv")

---
# End of notebook
