==> 08-anova_backtesting.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from scipy.stats import f_oneway, ttest_ind

plt.rcParams['figure.figsize'] = (10, 4)
%load_ext autoreload
%autoreload 2
font = {
    'weight' : 'regular',
    'family': 'serif',
    'size'   : 16
}
plt.rc('font', **font)

scores_lls = pd.read_excel(
    f"model_outputs/scores/lls_predictions-METEO-FRANCE_per_vineyard.xlsx"
)
scores_lls = scores_lls[scores_lls["Test variable"] == "Spearman"]
scores_model_pool = pd.read_excel(f"model_outputs/scores/model_pool_per_vineyard.xlsx")
scores_model_pool = scores_model_pool[scores_model_pool["Test variable"] == "Spearman"]

display(scores_lls.head())
display(scores_model_pool.head())

## ANOVA for phenology adapted features

sns.violinplot(data=scores_model_pool.drop(['Test variable'], axis=1))
plt.xticks(rotation=90)
plt.show()

print(f_oneway(
    scores_model_pool["OLS"],
    scores_model_pool["Ashenfelter OLS"],
    scores_model_pool["Quadratic OLS"],
))

print(ttest_ind(
    scores_model_pool["OLS"],
    scores_model_pool["Ashenfelter OLS"],
    equal_var=False
))

## Welch test for time-varying structure
Welch test can be used without any assumption about equality of variance (unlike the student t test)

sns.violinplot(data=scores_lls.drop(['Test variable'], axis=1))
plt.xticks(rotation=90)
plt.show()

ttest_ind(
    scores_lls["Classical OLS"],
    scores_lls["LLS gaussian kernel"],
    equal_var=False
)

---
## End of notebook
